+++ src/views/compare-view.tsx
@@ -143,14 +143,14 @@ export const CompareViewContainer = memo(
     }
 
     return (
-      <ThemeContext.Provider value={compareTheme}>
+      // <ThemeContext.Provider value={compareTheme}>
         <CompareView
           glCanvas={glCanvas}
           activeProfileState={activeProfileState}
           profileGroup={profileGroup}
           compareProfileGroup={compareProfileGroup}
         />
-      </ThemeContext.Provider>
+      // </ThemeContext.Provider>
     )
   },
 )
